requirements.txt
fastai==2.7.12
matplotlib==3.4.3
numpy==1.17.5
opencv_contrib_python==4.6.0.66
pandas==1.3.1
Pillow==9.5.0
scikit_image==0.19.3
scikit_learn==0.24.2
scipy==1.7.3
tifffile==2021.11.2
torch==1.10.1
tqdm==4.64.0
pylandsat==0.6.0
